Reset repo from origin

+++ basics/src/main/java/be/ecam/basics/exercises/Counter.java
@@ -1,8 +1,8 @@
 package be.ecam.basics.exercises;
 
 public class Counter {
-    public static int count(int start, int steps) {
-        int c = (int) start;
+    public static byte count(int start, int steps) {
+        byte c = (byte) start;
         for (int i = 0; i < steps; i++) {
             c++;
         }
